pyproject.toml


[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-compliance_env"
version = "0.1.0"
description = "Compliance Env environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "openai>=1.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "compliance_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["compliance_env", "compliance_env.server"]
package-dir = { "compliance_env" = ".", "compliance_env.server" = "server" }